final tweaks for 1st release

+++ GD4ClassCode/Aircraft.cpp
@@ -252,7 +252,7 @@ void Aircraft::setIdentifier(int identifier)
 	mIdentifier = identifier;
 }
 
-int	Aircraft::getKnockback()
+float Aircraft::getKnockback()
 {
 	return mKnockbackModifier;
 }
