+++ GD4ClassCode/Aircraft.hpp
@@ -42,7 +42,7 @@ public:
 	void					playLocalSound(CommandQueue& commands, SoundEffect::ID effect);
 	int						getIdentifier();
 	void					setIdentifier(int identifier);
-	int						getKnockback();
+	float						getKnockback();
 	void					setKnockback(float increment);
 	void					incrementKnockback(float increment);
 	int						getMissileAmmo() const;
